requirements.txt
Brotli==1.0.7
dj-config-url==0.1.1
Django==2.2.12
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1